+++ extensions/identity-hub/build.gradle.kts
@@ -27,7 +27,7 @@ val nimbusVersion: String by project
 
 dependencies {
     api(project(":spi:identity-hub-spi"))
-    api(project(":identity-hub-dtos"))
+    api(project(":identity-hub-model"))
     implementation(project(":spi:identity-hub-store-spi"))
     implementation("${edcGroup}:http:${edcVersion}")
     implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
