--- identity-hub-did/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-org.eclipse.dataspaceconnector.identityhub.did.IdentityHubDidExtension
