--- identity-hub-did/src/test/java/org/eclipse/dataspaceconnector/identityhub/did/IdentityHubCredentialsVerifierTest.java
@@ -1,167 +1,0 @@
-/*
- *  Copyright (c) 2022 Microsoft Corporation
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Microsoft Corporation - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.identityhub.did;
-
-import com.github.javafaker.Faker;
-import com.nimbusds.jose.JWSAlgorithm;
-import com.nimbusds.jose.JWSHeader;
-import com.nimbusds.jwt.JWTClaimsSet;
-import com.nimbusds.jwt.SignedJWT;
-import org.eclipse.dataspaceconnector.iam.did.spi.credentials.CredentialsVerifier;
-import org.eclipse.dataspaceconnector.iam.did.spi.document.DidDocument;
-import org.eclipse.dataspaceconnector.iam.did.spi.document.Service;
-import org.eclipse.dataspaceconnector.identityhub.client.IdentityHubClient;
-import org.eclipse.dataspaceconnector.spi.monitor.ConsoleMonitor;
-import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
-import org.eclipse.dataspaceconnector.spi.response.ResponseStatus;
-import org.eclipse.dataspaceconnector.spi.response.StatusResult;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.dataspaceconnector.identityhub.junit.testfixtures.VerifiableCredentialTestUtil.buildSignedJwt;
-import static org.eclipse.dataspaceconnector.identityhub.junit.testfixtures.VerifiableCredentialTestUtil.generateVerifiableCredential;
-import static org.eclipse.dataspaceconnector.identityhub.junit.testfixtures.VerifiableCredentialTestUtil.toMap;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-public class IdentityHubCredentialsVerifierTest {
-
-    private static final Faker FAKER = new Faker();
-    private static final Monitor MONITOR = new ConsoleMonitor();
-    private IdentityHubClient identityHubClient = mock(IdentityHubClient.class);
-    private JwtCredentialsVerifier jwtCredentialsVerifier = mock(JwtCredentialsVerifier.class);
-    private CredentialsVerifier credentialsVerifier = new IdentityHubCredentialsVerifier(identityHubClient, MONITOR, jwtCredentialsVerifier);
-    private String hubBaseUrl = "https://" + FAKER.internet().url();
-    String issuer = FAKER.internet().url();
-    String subject = FAKER.internet().url();
-
-    @Test
-    public void getVerifiedClaims_getValidClaims() throws Exception {
-
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance()
-                .service(List.of(new Service("IdentityHub", "IdentityHub", hubBaseUrl))).build();
-        var credential = generateVerifiableCredential();
-        var jws = buildSignedJwt(credential, issuer, subject);
-        setUpMocks(jws, true, true);
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.succeeded());
-        assertThat(credentials.getContent())
-                .usingRecursiveComparison()
-                .ignoringFields(String.format("%s.exp", credential.getId()))
-                .isEqualTo(toMap(credential, issuer, subject));
-    }
-
-    private void setUpMocks(SignedJWT jws, boolean isSigned, boolean claimsValid) {
-        when(identityHubClient.getVerifiableCredentials(hubBaseUrl)).thenReturn(StatusResult.success(List.of(jws)));
-        when(jwtCredentialsVerifier.isSignedByIssuer(jws)).thenReturn(isSigned);
-        when(jwtCredentialsVerifier.verifyClaims(eq(jws), any())).thenReturn(claimsValid);
-    }
-
-    @Test
-    public void getVerifiedClaims_filtersSignedByWrongIssuer() throws Exception {
-
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance()
-                .service(List.of(new Service("IdentityHub", "IdentityHub", hubBaseUrl))).build();
-        var credential = generateVerifiableCredential();
-        var jws = buildSignedJwt(credential, issuer, subject);
-        setUpMocks(jws, true, false);
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.succeeded());
-        assertThat(credentials.getContent().size()).isEqualTo(0);
-    }
-
-    @Test
-    public void getVerifiedClaims_hubUrlNotResolved() {
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance().build();
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.failed());
-    }
-
-    @Test
-    public void getVerifiedClaims_idHubCallFails() {
-
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance().service(List.of(new Service("IdentityHub", "IdentityHub", hubBaseUrl))).build();
-        when(identityHubClient.getVerifiableCredentials(hubBaseUrl)).thenReturn(StatusResult.failure(ResponseStatus.FATAL_ERROR));
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.failed());
-    }
-
-    @Test
-    public void getVerifiedClaims_verifiableCredentialsWithWrongFormat() {
-
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance()
-                .service(List.of(new Service("IdentityHub", "IdentityHub", hubBaseUrl))).build();
-        var jws = new SignedJWT(new JWSHeader.Builder(JWSAlgorithm.ES256).build(), new JWTClaimsSet.Builder().build());
-        setUpMocks(jws, true, true);
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.succeeded());
-        assertThat(credentials.getContent().isEmpty());
-    }
-
-    @Test
-    public void getVerifiedClaims_verifiableCredentialsWithMissingId() {
-
-        // Arrange
-        var didDocument = DidDocument.Builder.newInstance()
-                .service(List.of(new Service("IdentityHub", "IdentityHub", hubBaseUrl))).build();
-        var jwsHeader = new JWSHeader.Builder(JWSAlgorithm.ES256).build();
-        var jwtClaims = new JWTClaimsSet.Builder()
-                .claim("vc", Map.of(FAKER.lorem().word(), FAKER.lorem().word()))
-                .issuer(FAKER.lorem().word())
-                .subject(subject)
-                .build();
-        var jws = new SignedJWT(jwsHeader, jwtClaims);
-        setUpMocks(jws, true, true);
-
-        // Act
-        var credentials = credentialsVerifier.getVerifiedCredentials(didDocument);
-
-        // Assert
-        assertThat(credentials.succeeded());
-        assertThat(credentials.getContent().isEmpty());
-    }
-
-}
