+++ identity-hub-verifier/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,0 +1,1 @@
+org.eclipse.dataspaceconnector.identityhub.verifier.IdentityHubDidExtension
