+++ settings.gradle.kts
@@ -4,6 +4,6 @@ include(":spi:identity-hub-spi")
 include(":spi:identity-hub-store-spi")
 include(":extensions:identity-hub")
 include(":identity-hub-client")
-include(":identity-hub-dtos")
-include(":identity-hub-did")
+include(":identity-hub-model")
+include(":identity-hub-verifier")
 include(":identity-hub-junit")
